fix: Phase 1 compilation errors - align with actual model schemas
- Fixed destination_clusters.rs field names (hypervisor, status, network fields)
- Fixed hardware_pool.rs allocation fields (purpose, allocation_type, workflow_id)
- Fixed HardwarePool capacity field references (cpu_cores_total, optional fields)
- Fixed metadata types (HashMap<String, serde_json::Value>)
- Fixed overcommit ratio type casting (f32 vs f64)
- Added PartialEq, Eq, Hash derives to NetworkPurpose enum
- Handled all ValidationSeverity cases in match statements

All Phase 1 backend APIs now compile successfully.

+++ backend/src/services/capacity_planner_service.rs
@@ -268,8 +268,8 @@ impl CapacityPlannerService {
 
         // Calculate available capacity with overcommit
         let mut available_capacity = cluster.capacity_totals.clone();
-        available_capacity.cpu_cores = ((available_capacity.cpu_cores - ha_reserved.cpu_cores) as f64 * overcommit.cpu_ratio) as i32;
-        available_capacity.memory_gb = ((available_capacity.memory_gb - ha_reserved.memory_gb) as f64 * overcommit.memory_ratio) as i32;
+        available_capacity.cpu_cores = ((available_capacity.cpu_cores - ha_reserved.cpu_cores) as f32 * overcommit.cpu_ratio) as i32;
+        available_capacity.memory_gb = ((available_capacity.memory_gb - ha_reserved.memory_gb) as f32 * overcommit.memory_ratio) as i32;
         available_capacity.storage_gb -= ha_reserved.storage_gb;
 
         Ok(ClusterCapacityDetail {
